feat(backend): store money/price/quantity columns as exact NUMERIC
- add core.database.MONEY = Numeric(20, 8, asdecimal=False) and use it for the
  financial columns of Order, Portfolio, Holding and PriceAlert (replacing lossy
  Float/double precision); Python-side values stay float so trading arithmetic
  is unchanged and behaviour-preserving
- core/migrations.apply_type_migrations converts pre-existing PostgreSQL float
  columns to NUMERIC(20,8) in place (idempotent; inspects current type first)
- verified end-to-end on sqlite: fractional BUY/SELL, average-price accumulation
  and portfolio valuation all pass with the new column type

Note: full Python-side Decimal arithmetic (asdecimal=True) is a recommended
follow-up that requires a live-DB test pass.

+++ backend/models/order.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import DateTime, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from backend.core.database import Base
+from backend.core.database import Base, MONEY
 class Order(Base):
     __tablename__ = "orders"
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
@@ -10,12 +10,12 @@ class Order(Base):
     broker: Mapped[str] = mapped_column(String(50), nullable=False)
     ticker: Mapped[str] = mapped_column(String(50), nullable=False, index=True)
     action: Mapped[str] = mapped_column(String(10), nullable=False)
-    quantity_requested: Mapped[float] = mapped_column(Float, nullable=False)
-    quantity_filled: Mapped[float] = mapped_column(Float, default=0.0)
+    quantity_requested: Mapped[float] = mapped_column(MONEY, nullable=False)
+    quantity_filled: Mapped[float] = mapped_column(MONEY, default=0.0)
     status: Mapped[str] = mapped_column(String(30), default="PENDING")
-    price_per_share: Mapped[float | None] = mapped_column(Float, nullable=True)
-    total_value: Mapped[float | None] = mapped_column(Float, nullable=True)
-    commission: Mapped[float] = mapped_column(Float, default=0.0)
+    price_per_share: Mapped[float | None] = mapped_column(MONEY, nullable=True)
+    total_value: Mapped[float | None] = mapped_column(MONEY, nullable=True)
+    commission: Mapped[float] = mapped_column(MONEY, default=0.0)
     analysis_id: Mapped[int | None] = mapped_column(
         Integer, ForeignKey("analysis_results.id"), nullable=True
     )
